update parseArrayElement to use DeferRenderExpression
We were implicitly using LiteralExpression to defer rendering of templates set as values within ArrayExpressions created by array element tags. This didn't correctly capture the context at the time of evaluating the expression. Replace with the purpose-built DeferRenderExpression, which fixes this and also is more explicit in purpose.

Also, minor refactor of parseArrayElement, which can now appropriately distinguish based on the type of the expression returned, making the parsing code more explicit.